Move SaneAR and VMR9/EVR/Sync renderer settings into popup dialogs, redesign Output page (#3146)
