+++ src/mpc-hc/PPageAudioRenderer.cpp
@@ -28,8 +28,6 @@
 #include <Mmdeviceapi.h>
 #include <sanear/src/Settings.h>
 #include "AppSettings.h"
-#include "PPageOutput.h"
-#include "FGManager.h"
 
 namespace
 {
@@ -69,13 +67,12 @@ namespace
 
 IMPLEMENT_DYNAMIC(CPPageAudioRenderer, CMPCThemePPageBase)
 CPPageAudioRenderer::CPPageAudioRenderer()
-    : CMPCThemePPageBase(IDD, IDD_PPAGEAUDIORENDERER)
+    : CMPCThemePPageBase(IDD, IDS_PPAGE_AUDIORENDERER_TITLE)
     , m_bExclusiveMode(FALSE)
     , m_bCrossfeedEnabled(FALSE)
     , m_bIgnoreSystemChannelMixer(TRUE)
-    , m_bIsEnabled(FALSE)
-    , curAudioRenderer()
 {
+    m_bPopupHosted = true;
 }
 
 void CPPageAudioRenderer::DoDataExchange(CDataExchange* pDX)
@@ -84,8 +81,7 @@ void CPPageAudioRenderer::DoDataExchange(CDataExchange* pDX)
     DDX_Check(pDX, IDC_CHECK1, m_bExclusiveMode);
     DDX_Check(pDX, IDC_CHECK3, m_bCrossfeedEnabled);
     DDX_Check(pDX, IDC_CHECK4, m_bIgnoreSystemChannelMixer);
-    DDX_Check(pDX, IDC_CHECK5, m_bIsEnabled);
-    
+
     DDX_Control(pDX, IDC_COMBO1, m_combo1);
     DDX_Control(pDX, IDC_SLIDER1, m_slider1);
     DDX_Control(pDX, IDC_SLIDER2, m_slider2);
@@ -103,20 +99,6 @@ BEGIN_MESSAGE_MAP(CPPageAudioRenderer, CMPCThemePPageBase)
     ON_UPDATE_COMMAND_UI(IDC_STATIC3, OnUpdateCrossfeedGroup)
     ON_UPDATE_COMMAND_UI(IDC_SLIDER2, OnUpdateCrossfeedGroup)
     ON_UPDATE_COMMAND_UI(IDC_STATIC4, OnUpdateCrossfeedLevelLabel)
-
-    ON_BN_CLICKED(IDC_CHECK5, OnClickInternalAudioRenderer)
-    ON_UPDATE_COMMAND_UI(IDC_COMBO1, OnUpdateInternalAudioEnabled)
-    ON_UPDATE_COMMAND_UI(IDC_CHECK1, OnUpdateInternalAudioEnabled)
-    ON_UPDATE_COMMAND_UI(IDC_CHECK3, OnUpdateInternalAudioEnabled)
-    ON_UPDATE_COMMAND_UI(IDC_CHECK4, OnUpdateInternalAudioEnabled)
-    ON_UPDATE_COMMAND_UI(IDC_STATIC5, OnUpdateInternalAudioEnabled)
-    ON_UPDATE_COMMAND_UI(IDC_STATIC6, OnUpdateInternalAudioEnabled)
-    ON_UPDATE_COMMAND_UI(IDC_STATIC7, OnUpdateInternalAudioEnabled)
-    ON_UPDATE_COMMAND_UI(IDC_STATIC8, OnUpdateInternalAudioEnabled)
-
-    ON_UPDATE_COMMAND_UI(IDC_BUTTON3, OnUpdateMPCAudioRenderer)
-    ON_BN_CLICKED(IDC_BUTTON3, OnMPCAudioRendererButton)
-
 END_MESSAGE_MAP()
 
 BOOL CPPageAudioRenderer::OnInitDialog()
@@ -176,14 +158,6 @@ BOOL CPPageAudioRenderer::OnInitDialog()
     m_slider1.SetPos(crossfeedCuttoffFrequency);
     m_slider2.SetPos(crossfeedLevel);
 
-    CPPageOutput* po = static_cast<CPPageOutput*>(FindSiblingPage(RUNTIME_CLASS(CPPageOutput)));
-    if (po) { //output page visible, so we will use its setting (maybe unapplied)
-        curAudioRenderer = po->GetAudioRendererDisplayName();
-    } else {
-        curAudioRenderer = s.SelectedAudioRenderer();
-    }
-    m_bIsEnabled = (curAudioRenderer == AUDRNDT_INTERNAL);
-
     UpdateData(FALSE);
 
     return TRUE;
@@ -210,48 +184,14 @@ BOOL CPPageAudioRenderer::OnApply()
     s.sanear->SetCrossfeedSettings(m_slider1.GetPos(), m_slider2.GetPos());
     s.sanear->SetCrossfeedEnabled(m_bCrossfeedEnabled);
     s.sanear->SetIgnoreSystemChannelMixer(m_bIgnoreSystemChannelMixer);
-    AfxGetAppSettings().strAudioRendererDisplayName = curAudioRenderer;
 
     return __super::OnApply();
-}
-
-void CPPageAudioRenderer::OnCancel()
-{
-    __super::OnCancel();
 }
 
 void CPPageAudioRenderer::OnHScroll(UINT nSBCode, UINT nPos, CScrollBar* pScrollBar)
 {
     if (pScrollBar && (*pScrollBar == m_slider1 || *pScrollBar == m_slider2)) {
         SetModified(TRUE);
-    }
-}
-
-void CPPageAudioRenderer::OnClickInternalAudioRenderer()
-{
-    auto& s = AfxGetAppSettings();
-    UpdateData(TRUE);
-    if (m_bIsEnabled) { //we just enabled it, force to internal
-        curAudioRenderer = AUDRNDT_INTERNAL;
-    } else { //we just disabled it, force to last used
-        if (s.SelectedAudioRenderer() != AUDRNDT_INTERNAL) {
-            curAudioRenderer = s.SelectedAudioRenderer();
-        } else { //unless it was already set to internal before, now unchecking it means "use default"
-            curAudioRenderer = _T("");
-        }
-    }
-
-    CPPageOutput* po = static_cast<CPPageOutput*>(FindSiblingPage(RUNTIME_CLASS(CPPageOutput)));
-    if (po) { //output page visible, so we have to update the dropdown
-        po->UpdateAudioRenderer(curAudioRenderer);
-    }
-
-    SetModified(TRUE);
-}
-
-void CPPageAudioRenderer::OnMPCAudioRendererButton() {
-    if (curAudioRenderer == AUDRNDT_MPC) {
-        ShowPPage(CFGManager::GetMpcAudioRendererInstance);
     }
 }
 
@@ -271,25 +211,7 @@ void CPPageAudioRenderer::OnJMeierButton()
 
 void CPPageAudioRenderer::OnUpdateCrossfeedGroup(CCmdUI* pCmdUI)
 {
-    pCmdUI->Enable(IsDlgButtonChecked(IDC_CHECK3) && m_bIsEnabled);
-}
-
-void CPPageAudioRenderer::OnUpdateInternalAudioEnabled(CCmdUI* pCmdUI) {
-    pCmdUI->Enable(m_bIsEnabled);
-}
-
-void CPPageAudioRenderer::OnUpdateMPCAudioRenderer(CCmdUI* pCmdUI) {
-    pCmdUI->Enable(curAudioRenderer == AUDRNDT_MPC);
-}
-
-
-void CPPageAudioRenderer::SetEnabled(bool enabled) {
-    m_bIsEnabled = enabled;
-}
-
-void CPPageAudioRenderer::SetCurAudioRenderer(CString renderer) {
-    curAudioRenderer = renderer;
-    SetEnabled(renderer == AUDRNDT_INTERNAL);
+    pCmdUI->Enable(IsDlgButtonChecked(IDC_CHECK3));
 }
 
 void CPPageAudioRenderer::OnUpdateCrossfeedCutoffLabel(CCmdUI* pCmdUI)
