+++ src/mpc-hc/PPageAudioRenderer.h
@@ -33,28 +33,21 @@ public:
     CPPageAudioRenderer();
 
     enum { IDD = IDD_PPAGEAUDIORENDERER };
-    void SetEnabled(bool enabled);
-    void SetCurAudioRenderer(CString renderer);
-    void OnMPCAudioRendererButton();
 
 protected:
 
     void DoDataExchange(CDataExchange* pDX) override;
     BOOL OnInitDialog() override;
     BOOL OnApply() override;
-    void OnCancel() override;
 
     void OnHScroll(UINT nSBCode, UINT nPos, CScrollBar* pScrollBar);
 
     void OnCMoyButton();
     void OnJMeierButton();
-    void OnClickInternalAudioRenderer();
 
     void OnUpdateCrossfeedGroup(CCmdUI* pCmdUI);
     void OnUpdateCrossfeedCutoffLabel(CCmdUI* pCmdUI);
     void OnUpdateCrossfeedLevelLabel(CCmdUI* pCmdUI);
-    void OnUpdateInternalAudioEnabled(CCmdUI* pCmdUI);
-    void OnUpdateMPCAudioRenderer(CCmdUI* pCmdUI);
 
     DECLARE_MESSAGE_MAP()
 
@@ -63,8 +56,6 @@ protected:
     BOOL m_bExclusiveMode;
     BOOL m_bCrossfeedEnabled;
     BOOL m_bIgnoreSystemChannelMixer;
-    BOOL m_bIsEnabled;
-    CString curAudioRenderer;
 
     CMPCThemeComboBox m_combo1;
     CMPCThemeSliderCtrl m_slider1;
